Split up datatypes for formulas
The problem is, that the current type system mixes up typing which
is fixed in the program (as Bool for logic formualas) and typing
which is done in the formula itself (as types of variables etc.).

So now the types will be split into those parts and relations
which are in between as their parameters get their types in the
formulas and the result is Bool.