Changed strategy from type checking to type infering
Till now it has been assumed that one knows the type of
an expression. But e.g. for relations this is not the case:
in the expression "a = b" the types of a and b are not known.
So we have to infer the type for them first and are then able
to assemble it to an expression of typ Bool.